build
Modifies `applicationId`, `versionCode`, and `versionName` to allow this modified version to be installed side-by-side with the official whoBIRD.

"ssp." identifies this as a subspecies of whoBIRD, with "fgs" to indicate that it uses a foreground service.

versionCode `3801` indicates whoBIRD version `38` with "ssp. FGS" version `01`.
versionName `3.8.0.1` indicates whoBIRD `3.8` with "ssp. FGS" version `0.1`